having some issues getting php.ini to match craft requirements.
I have searched (find) for all the files matching 'php.ini' and change each one to match the requirements with no success. everything else is going smoothly. setting up sections and fields. everything is ready for making some entries. also it looks like I can set this project up with an angular front so that it is an SPA. I get everything going with the .twig convention. once it's looking good, I give it the sheen of an angular front end.