single mobile is near complete.
need to work out toggle action for the hamburger and menu. also forgot to add a muting overlay for the other pictures on the home page when one is selected. and eventually I would like to add some tooltips to help explain the footer icons and toggle action on the nav bar.